Add a responsive Header
